♻️ chore: clean code sections
