+++ src/components/general/app-sidebar.tsx
@@ -6,11 +6,11 @@ import {
   SidebarGroupLabel,
   SidebarHeader,
   SidebarMenu,
-  SidebarSeparator,
 } from '@/components/ui/sidebar';
 import { MenuAvatar } from '@/components/menu/menu-avatar/menu-avatar';
 import { MenuItems } from '@/components/menu/menu-items/menu-items';
 import { useTranslations } from 'next-intl';
+import { SocialItems } from '@/components/social/social-items';
 
 export const AppSidebar = () => {
   const t = useTranslations('Sidebar');
@@ -20,37 +20,18 @@ export const AppSidebar = () => {
       <SidebarHeader>
         <div className="flex flex-col place-items-center gap-y-2">
           <MenuAvatar role={t('role')} />
+          <SocialItems size="sm" />
         </div>
       </SidebarHeader>
       <SidebarContent className="justify-between">
-        <div>
-          <SidebarGroup>
-            <SidebarGroupLabel>{t('professional.title')}</SidebarGroupLabel>
-            <SidebarGroupContent>
-              <SidebarMenu>
-                <MenuItems items={t.raw('professional.links')} />
-              </SidebarMenu>
-            </SidebarGroupContent>
-          </SidebarGroup>
-          <SidebarGroup>
-            <SidebarGroupLabel>{t('hobbies.title')}</SidebarGroupLabel>
-            <SidebarGroupContent>
-              <SidebarMenu>
-                <MenuItems items={t.raw('hobbies.links')} />
-              </SidebarMenu>
-            </SidebarGroupContent>
-          </SidebarGroup>
-        </div>
-        <div>
-          <SidebarSeparator />
-          <SidebarGroup>
-            <SidebarGroupContent>
-              <SidebarMenu>
-                <MenuItems size="sm" items={t.raw('footer.links')} />
-              </SidebarMenu>
-            </SidebarGroupContent>
-          </SidebarGroup>
-        </div>
+        <SidebarGroup>
+          <SidebarGroupLabel>{t('professional.title')}</SidebarGroupLabel>
+          <SidebarGroupContent>
+            <SidebarMenu>
+              <MenuItems items={t.raw('professional.links')} />
+            </SidebarMenu>
+          </SidebarGroupContent>
+        </SidebarGroup>
       </SidebarContent>
     </Sidebar>
   );
